requirements.txt
arxiv==2.2.0
fitz==0.0.1.dev2
python-dotenv==1.1.0
python_docx==1.1.2
Requests==2.32.3




